+++ R/add1s_y2.R
@@ -3,13 +3,17 @@
 ### Description
 #' Add 1 Blank PowerPoint slide
 #'
-#' This function adds a new PowerPoint slide. For this function to work, you need a powerpoint object saved into R called "doc"
-#' @param title DEFAULT: "Title"; Add a slide title in quotes, automatically formatted to the Default Qualtrics template
-#' @param commentary DEFAULT: "Commentary"; Add commentary/desctiption in quotes, automatically formatted to the Default Qualtrics template
-#' @param footer DEFAULT: "Footer"; Add a footer in quotes
+#' This function adds a new PowerPoint slide. For this function to work, you need a powerpoint object saved into R called 'doc'
+#' @param title DEFAULT: 'Title'; Add a slide title in quotes, automatically formatted to the Default Qualtrics template
+#' @param commentary DEFAULT: 'Commentary'; Add commentary/desctiption in quotes, automatically formatted to the Default Qualtrics template
+#' @param footer DEFAULT: 'Footer'; Add a footer in quotes
 #' @param text_boxes DEFAULT = FALSE; Automatically adds title, commentary, and footer boxes to the blank slide. Set to F for no text boxes
-#' @param slide_name DEFAULT: "Findings / 1 chart"; The name of the type of the PP slide you want added to the PP
-#' @param master_name DEFAULT: "Office Theme"; The name of the PP master layout that the slide_name comes from
+#' @param slide_name DEFAULT: 'Findings / 1 chart'; The name of the type of the PP slide you want added to the PP
+#' @param master_name DEFAULT: 'Office Theme'; The name of the PP master layout that the slide_name comes from
+#' @param title_color DEFAULT: 'Black'; Color or hexcode for slide title
+#' @param commentary_color DEFAULT: 'Black'; Color or hexcode for slide commentary
+#' @param footer_color DEFAULT: 'Black'; Color or hexcode for slide footer
+#' @param font_family DEFAULT: 'BentonSans Regular'; font for slide title, commentary, and footer
 #' @keywords powerpoint slide
 #' @export
 #' @examples
@@ -32,8 +36,12 @@ add1s_y2 <- function(
   commentary = 'Commentary',
   footer = 'Footer',
   text_boxes = TRUE,
-  slide_name = "Blank",
-  master_name = "1_Office Theme"
+  slide_name = 'Blank',
+  master_name = '1_Office Theme',
+  title_color = 'Black',
+  commentary_color = 'Black',
+  footer_color = 'Black',
+  font_family = 'BentonSans Regular'
 ) {
   doc <- officer::add_slide(
     doc,
@@ -41,9 +49,9 @@ add1s_y2 <- function(
     master = master_name
   )
   if (text_boxes == TRUE) {
-    doc <- doc %>% add_title_findings(title)
-    doc <- doc %>% add_commentary_findings(commentary)
-    doc <- doc %>% add_footer_findings(footer)
+    doc <- doc %>% add_title_findings(title, title_color, font_family)
+    doc <- doc %>% add_commentary_findings(commentary, commentary_color, font_family)
+    doc <- doc %>% add_footer_findings(footer, footer_color, font_family)
   } else {
     doc <- doc
   }
@@ -55,16 +63,16 @@ add1s_y2 <- function(
 add_title_findings <- function(
   doc,
   text,
-  text_color = "Black",
+  title_color,
+  font_family,
   font_size = 32,
-  font_family = "BentonSans Regular",
   left = 0.5,
   top = 0.25,
   width = 10.5,
   height = .75
 ){
   properties <- officer::fp_text(
-    color = text_color,
+    color = title_color,
     font.size = font_size,
     font.family = font_family
   )
@@ -90,16 +98,16 @@ add_title_findings <- function(
 add_commentary_findings <- function(
   doc,
   text,
-  text_color = "Black",
+  commentary_color,
+  font_family,
   font_size = 14,
-  font_family = "BentonSans Regular",
   left = 0.5,
   top = 1,
   width = 10.5,
   height = .75
 ){
   properties <- officer::fp_text(
-    color = text_color,
+    color = commentary_color,
     font.size = font_size,
     font.family = font_family
   )
@@ -125,16 +133,16 @@ add_commentary_findings <- function(
 add_footer_findings <- function(
   doc,
   text,
-  text_color = "Black",
+  footer_color,
+  font_family,
   font_size = 10,
-  font_family = "BentonSans Regular",
   left = 0.5,
   top = 7.13,
   width = 10.5,
   height = .25
 ){
   properties <- officer::fp_text(
-    color = text_color,
+    color = footer_color,
     font.size = font_size,
     font.family = font_family
   )